feat:Add HXL Tags validation
